requirement.txt
streamlit
langchain
firebase_admin
PyPDF2
langchain_google_genai
google-generativeai
langchain_community
gtts
huggingface_hub